Use `Signable` for serialisation concerns.
We earlier introduced explicit serialisers, but we now have the
`Signable` associated type, which can be used for exactly this. The
intention here is that, for example, the serialisation of old cardano
types may be done over the `Decoded` class, thus ensuring that we use
the correct bytes carried in the annotation.