Rename and move some CBOR utilities
{encode,decode}{KnownUknown}CborDataItem gets renamed to
{encode,decode}NestedCbor{Bytes}.

The CRC utils are removed entirely.

The rationale in both cases is that this package should be for general
utils that we expect to reuse across other packages, not for things that
are quirks of the Byron encodings. In this case the "known"/"unknown"
pattern is exclusive to the Byron encodings. This is not a pattern we
wish to perpetuate in new code. The nested CBOR is still useful
elsewhere so we keep the code but under a more appropriate name.

The old name, and the CRC utils will be moved into the cardano-ledger
package directly, where their scope can be limited to the Byron era
encodings.