Remove Ord instances for crypto keys
The pattern that our crypto libraries use is that they deliberately do not
provide Ord instances for keys (verification or signing keys). Instead our
crypto library encourage the use of comparisons on key hashes.

This patch removes Ord instances for keys and signatures. It removes Eq
instances for signing keys. It keeps Eq instances for verification keys
and signaures.