Add `abstractSizeVKey` and `abstractSizeSig` to `DSIGNAlgorithm`
There is s need to calculate the abstract size of a transaction in Shelley. The
size of a transaction also includes the witnesses which in turn contain the
verification key and signature.

Adding the above functions to the class allows having an upper bound of the
actual size of signatures and verification keys. For the un-bounded bitsize RSA
implementation, this is not possible.