Add a new Seed abstraction
This is intended to replace the use of MonadRandom for key generation.
Key generation is special in that it needs proper entropy, and the
amount it needs it bounded.

This makes MonadRandom a poor match. It is all too easy to pick the wrong
instance for MonadRandom and end up with something that does not get
proper entropy. Also, MonadRandom is far too general in that it can get
arbitary amounts of data when key generation uses bounded entropy.

So the intention is to replace MonadRandom for key generation with an
explicit Seed argument, and make key generation pure. We then need to be
explicit about how much entropy we need to pass, since it must fail if
not enough is provided.