Do the MonadRandom/Seed API switch for the KES API too
Remove use of MonadRandom from the KES API. The sign and update are
simply required to be pure and no replacement args. This limits us to
KES algorithms that do not need extra randomness for these operations,
which is fine for the ones we are going to use.

The KES key gen gets an explicit Seed argument and a class method to
tell us the required seed size, just like we did for DSIGN.

The KES tests need a bit of a cleanup now for their use of seeds. This
patch just did the minimal necessary changes to the tests.