Adjust KES interface to be fixed duration only
Prior to this patch, the KES class only supported KES algorithms that
are runtime-paramaterised by their duration and does not support
fixed-duration algorithms.

With this patch, we switch the KES class to only support KES algorithms
that have a fixed duration, and no longer supports variable-duration
algorithms.

This corresponds to the kind of KES algorithm that we will actually use.

Add a method to the KES class:

  totalPeriodsKES :: proxy v -> Natural

And remove the duration :: Natural arg of genKeyKES and seedSizeKES.

Adjust the existing Mock and Simple instances to still be paramaterised
by the duration, but at the type level.

In the tests we instantiate it to 42. In principle we could still
generate a duration and use SomeNat to run tests with an arbitrary
duration. It is unclear if it is worth it however.