Make the VRF key gen pure, like the others
We cannot entirely eliminate the MonadRandom from the VRF API for now
because our mock instance does still use it. We cannot have it use a
pure PRNG because the VRF API for creating a VRF proof does not return
a new VRF secret key (where we could stash a PRNG state).