Extend DSIGN class with raw serialisation
Add rawSerialise/rawDeserialise* methods to the DSIGN type class.

We need both CBOR serialisation and raw serialisation. It is important
that we be able to get at the raw format of keys. This is important for
example for hashing verification keys: for simplicity and documentation
purposes the hash should be of the raw key with no extra tags. Doing
anything else is inexplicable.

Provide defaults for the raw versions in terms of the CBOR versions,
which helps for the mock instances.

Provide defaults for the CBOR versions in terms of the raw versions,
which helps for the real instances.

Provide a MINIMAL pragma which explains these recursive defaults.