Adjust updateKES so it takes the current period, not the target
This follows the approach in the MMM paper. All the KES functions take
the current period as an argument, even the update operation. The update
also only increments the period by one. Previously the operation allowed
any target evolution. The new approach is in keeping with the rest of
the design that follows the MMM approach: a raw interface that does not
track the current period and has to have it supplied as an argument.

A wrapper could easily provide tracking of the current period, and if
needed provide a multi-step update that evolves repeatedly to get to the
target period.