Move CBOR en/decode from class methods to normal functions
Now that we have fixed-size serialisation for all the classes and
instances, we can do the CBOR wrappers in a uniform way.

Move the various encode/decode functions out of the classes, and make
them standalone functions. This means we remove the default
implementations for the raw ones: the raw versions must be provided.
This also mean no more MINIMAL pragma needed.

Adjust the To/FromCBOR instances to use the provided functions and not
do their own encoding. We have tests that check that these match up.

Overall this is just a simplification. We already require the raw
serialisation forms, so it's simpler to not allow the CBOR variants to
be overridden.