Add FiniteBytes
This can be used as a value secure memory foreign pointer points
to. Later, we might want to revisit this
as all sizes are static (hash digest, key length, ...)
and not use Storable at all. Additional benefit
would be from not having `peek` and `poke`.

But currently, being able to peek and show maybe useful.