Make OutputVRF a newtype and provide Natural conversions
OutputVRF contains bytes. Previously it was a type alias for ByteString.
This makes it a proper newtype (with a phantom type param to match the
type of key, etc).

We provide an access function for getting the VRF bytes, and a
conversion into Natural, since we need that for our use cases.