Add an option to use bundled VRF C code, with plain libsodium
There are two options for how to get the VRF C code:

1. Use a fork of libsodium containing the VRF code
2. Use an upstream libsodium and bundled VRF C code

This patch adds the second option. It has the downside that it
duplicates the ed25519 "ref10" implementation (which is contained in
libsodium but not exported via public symbols). It has the convenience
advantage of working with the version of libsodium available on most
Linux distributions.