Change Signable constraint to use SignableRepresentation
The approach of having the Signable constraint set to ToCBOR leads to
the mistake of signing and verifying the serialisation of a bytestring
rather than the bytestring itself. For example it leads to accidentally
signing the CBOR serialisation a txbody hash, rather than simply the
txbody hash.

So switch from ToCBOR to a new

class SignableRepresentation a where
    getSignableRepresentation :: a -> ByteString

Which just returns the representation as bytes for the purpose of
signing or verifying.