Prevent accidental reintroduction of Ord instances for keys
The auditors have recommended in the past to remove `Ord` instances for signing
and verification keys. Instead, the comparison should be done on the hashes of
the keys.

See:
* https://github.com/input-output-hk/cardano-ledger/pull/587
* https://github.com/input-output-hk/cardano-base/pull/38
* https://github.com/input-output-hk/cardano-base/pull/60

After having been removed, these instances have been accidentally reintroduced
and people have wondered a few times why the instances were missing.

Using `TypeError` as a superclass constraint for a blanket `Ord` instance, it is
no longer possible to *accidentally* reintroduce an `Ord` instance because of
the overlap. Not even as orphans.

When `Ord` is needed for testing purposes, one should use `Ord` on the hashes or
use a newtype.