Add Cardano.Slotting.Time types for slot length and important times
These are thin wrappers around `UTCTime` and `NominalDiffTime`. This module is
being copied almost exactly from
`Ouroboros.Consensus.BlockchainTime.WallClock.Types`.

This prepares for the Consensus Layer being able to inform the Ledger how to
convert slots to the times that the Plutus VM expects for a tx's validity
interval.