Ed25519: Simplify (and correct) deriveVerKeyDSIGN
Previously `deriveVerKeyDSIGN` worked by:
	- Extracting the seed from the signing key.
	- (Re-)generating the secret key (seed, ver key) from the seed.
	- Returning the verification key.

Not only is this unnecessary (since the signing key already contains the
verification key) but it was subject to a concurrency bug, since it
internally mutates the secret key pointer which we pass in.

Instead, we simply extract the relevant part from the provided signing
key.