Add `unsafeLinearExtendEpochInfo`.
This commit adds a function which can extend an `EpochInfo` linearly
given a starting point. This is needed to put in place a patch for a bug
that arose in the Alonzo era, where the `EpochInfo` was too generous in
translating slots.

In order to support this, this adds an additional function
`epochInfoSlotLength`, which returns the 'SlotLength' for the given
slot.