Serialization secp (#289)
Fix serialization and bindings to `secp256k1` library.

This PR addresses several security issues regarding 'internal' versus 'external' representations of verification keys and signatures for SECP256k1 curve signature algorithms. The original problem was that the `secp256k1` library (and by extension, the `secp256k1-haskell` library wrapping it) does not expect that verification keys (and in the case of ECDSA, signatures) are exposed as their direct binaries (the 'internal' representation): instead, you are supposed to use functions that perform serialization (producing the 'external' representation). This is a problem that `secp256k1-haskell` exacerbates, as it provides `Serialise` implementations for `PubKey` and `Sig`, despite those being security risks, as they both dump the internal representations _directly_.

This commit fixes the following:

* The serialization and deserialization methods of the `DSIGN` interface for both ECDSA and Schnorr signature verification algorithms have been updated to produce and consume 'external' forms for public keys.
* The serialization and deserialization methods of the ECDSA implementation for signatures have also been updated to produce and consume 'external' forms.
* The `secp256k1-haskell` library has been completely eliminated: it doesn't provide the proper serialization methods required for ECDSA signatures and public keys, and mixing it and the FFI is flaky and more complex than it needs to be. Plus, one less dependency, since we already have to FFI into `secp256k1` anyway.

Due to the elimination of `secp256k1-haskell`, the ECDSA implementation also has to gain a 'wrapper newtype' for hashed messages. I've also added a helper to safely produce these using any `HashAlgorithm` that produces 32-byte outputs: I believe that currently, only SHA256 applies, but it gives us forward compatibility in case we want it.

Step-by-step:
* Fix serialization for Schnorr verification keys
* Move SECP context out of Schnorr module
* Rename C constants to mark them as Schnorr-related
* ECDSA SECP256k1 port to FFI, remove haskell-secp256k1 dep
* Fix C binding types, remove commented-out code
* Switch to `ccall unsafe` for functions
* Better SECP256k1 error messages
* Deprecate `psbZero` and `psbFromBytes`, new PSB functionality
* Describe pkgconfig overrides
* Fix Nix comments
* Add global initialization routine: `Cardano.Crypto.Init` that initializes `libsodium` and `secp256k1`
* Fix SECP context initialization to use a `ForeignPtr` with finalizer
* Document use of `nullPtr` on `secpEcdsaSign`
* Add note about hashing ECDSA SECP256k1 signed messages yourself

Co-authored-by: Koz Ross <koz.ross@retro-freedom.nz>
Co-authored-by: Koz Ross <koz.ross@runbox.com>
Co-authored-by: Michael Peyton Jones <me@michaelpj.com>