Remove SafePinned.
SafePinned was used as a safety guard, mainly to aid debugging, which
would use an MVar to avoid use-after-free on unmanaged memory (in
practice, we used this on mlocked memory).

However, generalizing SafePinned using MonadMVar means that the
SafePinned now has to be parametrized over the MonadMVar instance
(because MVar is also parametrized now), and that parametrization then
needs to be bubbled up into the KES class, which would make it a
terribly leaky abstraction.

Besides, we were only using SafePinned on some parts of the SignKeySumKES (the
"r" fields), but not others (the "sk" fields), and it actually
introduces some runtime overhead for relatively meager gain.