Pass mlocking allocator around as an extra parameter
This allows us to inject mlock allocation logging for testing purposes,
without needing a MonadMLock typeclass, and without making the mlocking
primitives typeclass methods. Downside is that there's some explicit
parameter passing along the chain, but user code should be mostly
oblivious of this, as we provide API-compatible wrappers that use the
default implementation (mlockedMalloc).