Move `blst_util.h` from `install-includes` to `extra-source-files`
This avoids using the deprecated `includes` field and is more consistent
with what we do in other packages. The only reason to put it in
`install-includes` is for client code to be able to use it, but we want
client code to use the Haskell bindings and not the C bindings.