Switch to decoing `ByteString` first for `PackedBytes`
As it turns out `cborg` library whenever decoing `ByteString` it will
take a slice in constant time of the original large buffer that is
being fed to the decoder, while it can't do the same when decoding
`ByteArray#`. Concidering that for `PackedBytes` this intermediate
buffer is short lived, because it is only needed for packing operation,
it totally makes sense to not make a redundant copy of this buffer.